Check for raster width less than array footprint
In this case do one row and set the "dy" value to the footprint
so that the "multiscan" factor is one.  It might be better to directly
calculate the "multiscan" factor rather than passing "dy" to the calculation
routines.

This is an alternative to the restrictions imposed by "HITEC": its
JavaScript interface tried to prevent the entry of map sizes below
116x116 for HARP and restricted "dy" to only 116.4 if the map size
was less than 120.