Set a minimum integration time per point of 0.1s
Since rounding of the integration time has been removed, we need
to check that it does not go below the minimum possible value.

The alternative would be to force the time to 0.1s if a lower
value was calculated.  For now,  however, just raise an error if this
happens.  This has the advantage that successful calculations should
be reversible.