model: Add script to compare data to ITC results
This reads the combined Trx and Tsys data files and writes corresponding
files to another directory which include the calculated Trx and Tsys values.

Sky efficiency (i.e. transmission) was added as an additional
"extra output" of the ITC's _calculate_t_sys method so that it can
be included in the comparison files.