Handle the two parts of basket weave separately
Rather than taking the results for the complete calculation in
each direction and averaging them, attempt to consider the two
parts in a more separate manner.  This changes how various
results are accumulated: RMS values are combined in quadrature,
integration times are returned separately, and elapsed times
are added.

Note: where previously an extra output value "int_time" was returned,
in the basket weaving case this may now instead be "int_time_1"
and "int_time_2" (which can be added for an approximate total
integration time).