Add LO frequency limits for ʻAlaʻihi
This adds two new parameters to the "receiver_info.json" file:
f_lo_min and f_lo_max.  If these are present then, instead of the
previous check of being close to the end of the receiver's tuning
range, the LO frequency is compared to the limits.  This allows
a wider tuning range to be specified for ʻAlaʻihi, but accessing
it will require adjustment of the IF.