Move assignment of "eta_sky" extra output
Remove "eta_sky" from the "_get_efficiencies_temperatures" extra output
since it is a main output of this method.  Instead add it in the
"_calculate_t_sys" and "_estimate_t_rx_from_t_sys" methods.
This latter method was not included in the test suite, so add
it to the calculation test.